Move existing rustAnalyzer.ts under dedicated submodule
