+++ src/rust-analyzer/index.ts
@@ -1,0 +1,6 @@
+export {
+  createLanguageClient,
+  getServer,
+  setupClient,
+  setupProgress,
+} from './rustAnalyzer';
